refactor(VMenu): merge mixins into main menu
