+++ packages/vuetify/src/components/VMenu/VMenu.js
@@ -11,18 +11,12 @@ import Returnable from '../../mixins/returnable'
 import Toggleable from '../../mixins/toggleable'
 import Themeable from '../../mixins/themeable'
 
-// Component level mixins
-import Activator from './mixins/menu-activator'
-import Generators from './mixins/menu-generators'
-import Keyable from './mixins/menu-keyable'
-import Position from './mixins/menu-position'
-
 // Directives
 import ClickOutside from '../../directives/click-outside'
 import Resize from '../../directives/resize'
 
 // Helpers
-import { convertToUnit } from '../../util/helpers'
+import { convertToUnit, keyCodes } from '../../util/helpers'
 import ThemeProvider from '../../util/ThemeProvider'
 
 /* @vue/component */
@@ -42,14 +36,10 @@ export default Vue.extend({
   },
 
   mixins: [
-    Activator,
     Dependent,
     Delayable,
     Detachable,
-    Generators,
-    Keyable,
     Menuable,
-    Position,
     Returnable,
     Toggleable,
     Themeable
@@ -66,6 +56,7 @@ export default Vue.extend({
       default: true
     },
     disabled: Boolean,
+    disableKeys: Boolean,
     fullWidth: Boolean,
     maxHeight: { default: 'auto' },
     openOnClick: {
@@ -87,9 +78,12 @@ export default Vue.extend({
 
   data () {
     return {
+      calculatedTopAuto: 0,
       defaultOffset: 8,
       hasJustFocused: false,
-      resizeTimeout: null
+      listIndex: -1,
+      resizeTimeout: null,
+      tiles: []
     }
   },
 
@@ -151,8 +145,21 @@ export default Vue.extend({
   },
 
   watch: {
+    isActive (val) {
+      if (!val) this.listIndex = -1
+    },
     isContentActive (val) {
       this.hasJustFocused = val
+    },
+    listIndex (next, prev) {
+      if (next in this.tiles) {
+        const tile = this.tiles[next]
+        tile.classList.add('v-list-item--highlighted')
+        this.$refs.content.scrollTop = tile.offsetTop - tile.clientHeight
+      }
+
+      prev in this.tiles &&
+        this.tiles[prev].classList.remove('v-list-item--highlighted')
     }
   },
 
@@ -179,10 +186,153 @@ export default Vue.extend({
         })
       })
     },
+    calcScrollPosition () {
+      const $el = this.$refs.content
+      const activeTile = $el.querySelector('.v-list-item--active')
+      const maxScrollTop = $el.scrollHeight - $el.offsetHeight
+
+      return activeTile
+        ? Math.min(maxScrollTop, Math.max(0, activeTile.offsetTop - $el.offsetHeight / 2 + activeTile.offsetHeight / 2))
+        : $el.scrollTop
+    },
+    calcLeftAuto () {
+      if (this.isAttached) return 0
+
+      return parseInt(this.dimensions.activator.left - this.defaultOffset * 2)
+    },
+    calcTopAuto () {
+      const $el = this.$refs.content
+      const activeTile = $el.querySelector('.v-list-item--active')
+
+      if (!activeTile) {
+        this.selectedIndex = null
+      }
+
+      if (this.offsetY || !activeTile) {
+        return this.computedTop
+      }
+
+      this.selectedIndex = Array.from(this.tiles).indexOf(activeTile)
+
+      const tileDistanceFromMenuTop = activeTile.offsetTop - this.calcScrollPosition()
+      const firstTileOffsetTop = $el.querySelector('.v-list-item').offsetTop
+
+      return this.computedTop - tileDistanceFromMenuTop - firstTileOffsetTop - 1
+    },
+    changeListIndex (e) {
+      // For infinite scroll and autocomplete, re-evaluate children
+      this.getTiles()
+
+      if (e.keyCode === keyCodes.down && this.listIndex < this.tiles.length - 1) {
+        this.listIndex++
+        // Allow user to set listIndex to -1 so
+        // that the list can be un-highlighted
+      } else if (e.keyCode === keyCodes.up && this.listIndex > -1) {
+        this.listIndex--
+      } else if (e.keyCode === keyCodes.enter && this.listIndex !== -1) {
+        this.tiles[this.listIndex].click()
+      } else { return }
+      // One of the conditions was met, prevent default action (#2988)
+      e.preventDefault()
+    },
     closeConditional (e) {
       return this.isActive &&
         this.closeOnClick &&
         !this.$refs.content.contains(e.target)
+    },
+    genTransition () {
+      if (!this.transition) return this.genContent()
+
+      return this.$createElement('transition', {
+        props: {
+          name: this.transition
+        }
+      }, [this.genContent()])
+    },
+    genDirectives () {
+      // Do not add click outside for hover menu
+      const directives = !this.openOnHover && this.closeOnClick ? [{
+        name: 'click-outside',
+        value: () => { this.isActive = false },
+        args: {
+          closeConditional: this.closeConditional,
+          include: () => [this.$el, ...this.getOpenDependentElements()]
+        }
+      }] : []
+
+      directives.push({
+        name: 'show',
+        value: this.isContentActive
+      })
+
+      return directives
+    },
+    genContent () {
+      const options = {
+        attrs: this.getScopeIdAttrs(),
+        staticClass: 'v-menu__content',
+        'class': {
+          ...this.rootThemeClasses,
+          'v-menu__content--auto': this.auto,
+          'v-menu__content--fixed': this.activatorFixed,
+          'menuable__content__active': this.isActive,
+          [this.contentClass.trim()]: true
+        },
+        style: this.styles,
+        directives: this.genDirectives(),
+        ref: 'content',
+        on: {
+          click: e => {
+            e.stopPropagation()
+            if (e.target.getAttribute('disabled')) return
+            if (this.closeOnContentClick) this.isActive = false
+          },
+          keydown: this.onKeyDown
+        }
+      }
+
+      !this.disabled && this.openOnHover && (options.on.mouseenter = this.mouseEnterHandler)
+      this.openOnHover && (options.on.mouseleave = this.mouseLeaveHandler)
+
+      return this.$createElement('div', options, this.showLazyContent(this.$slots.default))
+    },
+    getTiles () {
+      this.tiles = this.$refs.content.querySelectorAll('.v-list-item')
+    },
+    mouseEnterHandler () {
+      this.runDelay('open', () => {
+        if (this.hasJustFocused) return
+
+        this.hasJustFocused = true
+        this.isActive = true
+      })
+    },
+    mouseLeaveHandler (e) {
+      // Prevent accidental re-activation
+      this.runDelay('close', () => {
+        if (this.$refs.content.contains(e.relatedTarget)) return
+
+        requestAnimationFrame(() => {
+          this.isActive = false
+          this.callDeactivate()
+        })
+      })
+    },
+    onKeyDown (e) {
+      if (e.keyCode === keyCodes.esc) {
+        // Wait for dependent elements to close first
+        setTimeout(() => { this.isActive = false })
+        const activator = this.getActivator()
+        this.$nextTick(() => activator && activator.focus())
+      } else if (e.keyCode === keyCodes.tab) {
+        setTimeout(() => {
+          if (!this.$refs.content.contains(document.activeElement)) {
+            this.isActive = false
+          }
+        })
+      } else {
+        this.changeListIndex(e)
+      }
     },
     onResize () {
       if (!this.isActive) return
